Refactor network into single-class modules
