+++ jarvis/network/__init__.py
@@ -1,5 +1,13 @@
 """Agent network package."""
 
-from .core import AgentNetwork
+from .agent_network import AgentNetwork
+from .base_agent import NetworkAgent
+from .message import Message
+from .capability import Capability
 
-__all__ = ["AgentNetwork"]
+__all__ = [
+    "AgentNetwork",
+    "NetworkAgent",
+    "Message",
+    "Capability",
+]
