+++ jarvis/network/message.py
@@ -1,0 +1,20 @@
+from __future__ import annotations
+
+import uuid
+from dataclasses import dataclass, field
+from datetime import datetime, timezone
+from typing import Any, Optional
+
+
+@dataclass
+class Message:
+    """Message passed between agents."""
+
+    id: str = field(default_factory=lambda: str(uuid.uuid4()))
+    from_agent: str = ""
+    to_agent: Optional[str] = None  # None = broadcast
+    message_type: str = ""
+    content: Any = None
+    request_id: str = ""
+    timestamp: datetime = field(default_factory=lambda: datetime.now(timezone.utc))
+    reply_to: Optional[str] = None  # For response tracking
